+++ Source/Ejecta/EJUtils/EJBindingTouchInput.h
@@ -9,7 +9,7 @@
 	JSStringRef jsIdentifierName, jsPageXName, jsPageYName, jsClientXName, jsClientYName;
 	JSObjectRef jsRemainingTouches, jsChangedTouches;
 	JSObjectRef jsTouchesPool[EJ_TOUCH_INPUT_MAX_TOUCHES];
-	int touchesInPool;
+	NSUInteger touchesInPool;
 }
 
 - (void)triggerEvent:(NSString *)name all:(NSSet *)all changed:(NSSet *)changed remaining:(NSSet *)remaining;
